Moved TouchKit Annotation => org.vaadin.spring.touchkit.annotation
